Test re-organization because we need lot of tests for oneshot
