+++ tests/integration_impl/mod.rs
@@ -1,0 +1,8 @@
+//  \ O /
+//  / * \    aiur: the homeplanet for the famous executors
+// |' | '|   (c) 2020 - present, Vladimir Zvezda
+//   / \
+
+pub mod oneshot_t;
+pub mod spawn_t;
+pub mod future_utils;
